+++ lib/src/extensions/build_context_extension.dart
@@ -1,5 +1,7 @@
 import "package:flutter/material.dart";
 
+import "string_extension.dart";
+
 extension BuildContextExtension on BuildContext {
   /// returns the current scaled text height for [fontSize]
   ///
@@ -25,5 +27,50 @@ extension BuildContextExtension on BuildContext {
     fS = scaler.scale(fS);
 
     return round ? fS.roundToDouble() : fS;
+  }
+
+  /// Creates a [TextPainter] for reuse
+  ///
+  /// use [StringExtension.measure] to extract useful information.
+  ///
+  /// must be disposed after usage
+  TextPainter setupTextPainter({
+    required String text,
+    BoxConstraints? constraints,
+    TextDirection? direction,
+    TextScaler? textScaler,
+    Locale? locale,
+    TextStyle? style,
+    TextAlign? textAlign,
+    int? maxLines,
+    String? ellipsis,
+    StrutStyle? strutStyle,
+    TextWidthBasis? textWidthBasis,
+    TextHeightBehavior? textHeightBehavior,
+  }) {
+    constraints ??= const BoxConstraints.expand();
+
+    direction ??= Directionality.maybeOf(this);
+    textScaler ??= MediaQuery.textScalerOf(this);
+    locale ??= Localizations.maybeLocaleOf(this);
+    style ??= DefaultTextStyle.of(this).style;
+
+    textAlign ??= TextAlign.start;
+    textWidthBasis ??= TextWidthBasis.parent;
+
+    TextPainter painter = TextPainter(
+      textDirection: direction,
+      textAlign: textAlign,
+      text: TextSpan(text: text, style: style, locale: locale),
+      maxLines: maxLines,
+      ellipsis: ellipsis,
+      locale: locale,
+      strutStyle: strutStyle,
+      textWidthBasis: textWidthBasis,
+      textHeightBehavior: textHeightBehavior,
+      textScaler: textScaler,
+    );
+
+    return painter;
   }
 }