+++ lib/ux_improvements.dart
@@ -8,6 +8,7 @@ export "src/adaptive_text_scaler/adaptive_text_scale.dart";
 export "src/adaptive_text_scaler/chained_text_scaler.dart";
 export "src/cloak/cloak.dart";
 export "src/extensions/build_context_extension.dart";
+export "src/extensions/string_extension.dart";
 export "src/fakeloading/fakeloading_widget.dart";
 export "src/glass/glass.dart";
 export "src/glass/glass_card.dart";