+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.14.3
+
+## Added ##
+
+* BuildContextExtension.setupTextPainter() to get a reusable Textpainter (For people who have an actual need to measure text sizes)
+* StringExtension.measure() which uses a Textpainter and updates it when necessary.
+
 ## 0.14.2
 
 ## Fixed ##